Add HOLD_ON_OTHER_KEY_PRESS option for tap-hold keys
Implement an additional option for tap-hold keys which converts the
tap-hold key press into a hold action immediately when another key is
pressed (this is different from the existing PERMISSIVE_HOLD option,
which selects the hold action when another key is tapped (pressed and
then released) while the tap-hold key is pressed).  The Mod-Tap keys
already behave in a similar way, unless the IGNORE_MOD_TAP_INTERRUPT
option is enabled (but with some additional delays); the added option
makes this behavior available for all other kinds of hold-tap keys.