id80: Enable LTO by default
Enabling LTO provides not only a noticeable size reduction (for one
particular testing keymap the size changed from 26282 bytes (91% flash)
to 23870 bytes (83% flash)), but also a significant speed improvement.
For the older version of the `id80` code with the ROW2COL matrix the
scan rate increased from 3989 to 4635 scans per second (16% increase).
For the current version of the `id80` code with the COL2ROW matrix the
scan rate increased from 5932 to 7850 scans per second (32% increase).