[keymap] id80:sigprof-test: Add dynamic macro support
Use the two remaining unassigned keys on the _FN layer (T and G) for
dynamic macros.  Actions to start recording of macros are assigned to
the same keys with the left Alt modifier (this is done by changing the
left Alt key into `LM(_FN_ALT, MOD_LALT)`).  The “stop recording” action
is assigned to Fn+Alt+Esc.