Make One Shot Layer clearing configurable
PR #8832 changed the behavior of One Shot Layer - now the layer is
cleared after any keypress event for which process_record_quantum()
returns false for any reason.  However, some people consider this new
behavior to be a bug (#9521).  Revert this part of changes made in #8832
and make clearing of One Shot Layer configurable.