Make GPIO functions for AVR atomic
On AVR MCUs the GPIO macros like writePinLow() perform read-modify-write
operations on PORTx and DDRx registers, unless their argument is a
compile time constant (in that case the compiler can optimize those
operations into single `cbi` or `sbi` instructions).  Usually this works
without any problems; however, the software PWM emulation code in
quantum/backlight/backlight_avr.c manipulates GPIOs from interrupt
handlers, and this may cause issues which manifest in a flickering
backlight on some boards (#5953).  The problem happens when the
read-modify-write sequence performed by the matrix scanning code is
interrupted by the software PWM handler, and the software PWM uses a pin
from the same port as one of the matrix row pins (or column pins for a
ROW2COL matrix); in this case the resumed read-modify-write code will
overwrite the update performed by the software PWM emulation code.

Change the implementation of all GPIO functions so that they are atomic
with respect to interrupts; this is achieved by disabling interrupts
around the read-modify-write operations.  This fixes #5953, and should
not break other boards, although there will be some additional overhead
for some GPIO operations (but the most common case of performing
writePinLow() or writePinHigh() on pins which are known at compile time
should have no additional overhead, because in this case the assembly
implementation using `cbi` or `sbi` should be picked by the compiler).