WIP: Rewrite GPIO macros with asm
Using the "Q" constraint does not work in some cases, and using the "m"
constraint is impossible due to different instruction names for access
with and without displacement.  Try using the "e" constraint with
predecrement and postincrement addressing to access the DDRx register,
while always loading the address of the PORTx register into the pointer
register pair.