WSL: Allow setup with qmk_firmware under /home
WSL 2 is almost unusable if the working tree is on the Windows
filesystem; however, QMK setup scripts for WSL tried to require
exactly that, apparently because it was assumed that Windows
executables located in the Linux filesystem cannot be run.  However,
at least in Windows 10 1909 the WSL subsystem can run Windows *.exe
files even if they are stored inside the Linux filesystem; the only
restriction is that those files must also be made executable in
Linux (`chmod +x`).  Remove the code which was blocking the setup if
the qmk_firmware directory was not under `/mnt`.