WSL: Avoid symlinking `~/qmk_firmware` if the repo is already there
The WSL setup script always offered to create the `~/qmk_firmware`
symlink to the real QMK repository location, because it assumed that the
repository is in some other place under `/mnt`.  However, with WSL 2 the
repository must be stored in the Linux filesystem (most likely inside
`/home`) to get usable performance, therefore storing the repository
directly in `~/qmk_firmware` is possible.  Do not offer to create the
`~/qmk_firmware` symlink in this case.