WSL: Fix running `qmk_driver_installer.exe` from Linux FS
The `cmd.exe` invocation used to run `qmk_driver_installer.exe` with
elevated rights did not work if the executable was in the Linux
filesystem, because the current directory in that case was converted to an
UNC path (`\\wsl$\linux-distro-name\home\...`), and CMD just refuses to
run if the current directory has an UNC path.  Use `powershell.exe`
instead if running in WSL.

Also make sure that `qmk_driver_installer.exe` has the execute permission
in Linux before trying to run it.