Do not install most flashing tools on WSL
Linux versions of flashing tools won't work in WSL, so it may be better
to skip installing them to avoid confusion when someone tries to use
them and wonders why they don't work.

However, the `dfu-util` package still needs to be installed, because
that package also contain the `dfu-suffix` utility, which is used for
building the firmware.