[keymap] xd87:sigprof-test: Use backslash as numpad layer switch on hold
Because there are no really good keys remaining free that could be used
as a numpad layer switch, use the backslash key for this purpose.  This
also requires enabling the `HOLD_ON_OTHER_KEY_PRESS_PER_KEY` and
`PERMISSIVE_HOLD_PER_KEY` options, because different LT keys now require
different tuning (the “hold on other key press” behavior is more
appropriate for the Caps Lock and Menu keys, while the backslash key can
be used in the middle of text, therefore “permissive hold” should be a
better option for that key).