[keymap] id80:sigprof-test: Replace Fn+Alt with Fn+Ctrl
Using the Fn+Alt layer turned out to be problematic, becase the Alt
keypress is still seen by the host, and if only some key combinations
which are internally handled by the keyboard (such as the macro
recording keys) are pressed before releasing Alt, this results in the
host seeing a tap on Alt, which opens the current window menu on
Windows.  Use Fn+Ctrl instead (a tap on Ctrl is ignored in all GUI
environments that I know of).