[WIP] WS2812 gpiodma: Add support for P2M DMA transfers
The errata sheet for STM32F411 chips has a section which might apply to
the DMA usage in the `gpiodma` driver (2.1.10: In some specific cases,
DMA2 data corruption occurs when managing AHB and APB2 peripherals in a
concurrent way).  The suggested workaround is to use the DMA2 memory
port for writing to GPIO registers.

Add the `WS2812_GPIODMA_PERIPHERAL_TO_MEMORY` option, which should be
defined for affected devices, and implement the code to set up DMA in a
peripheral-to-memory configuration.