WIP: Make RETRO_TAPPING more robust for multiple dual-role keys
The RETRO_TAPPING code did not work properly when multiple dual-role
keys were involved.  E.g., if A and B are both dual-role keys with
RETRO_TAPPING enabled, the following sequences did not give the expected
results:

1) Press A, press B, release B after tapping term - in this case
   RETRO_TAPPING did not work for B, because `retro_tapping_counter` was
   incremented for both A and B events.

2) Press A, tap X (a normal key), press B, release A - in this case
   RETRO_TAPPING was activated for A, because `retro_tapping_counter`
   was reset after the X tap, and then “press B, release A” counted as 2
   events, even though they were for different keys.

Instead of using just a counter, remember the matrix position of the
last key press event, and check whether the key release event actually
matches the key press event that came before it.  This test must be done
in `action_exec()`, because `process_action()` may handle events in a
different order if dual-role keys are involved (when the tap/hold
decision for a dual-role key is in progress, other key press events are
queued, but some key release events may still be handled immediately).
The `tap.retro_tapping` flag is then passed through `process_tapping()`
to `process_action()`, where it is used to trigger the retro tapping
behavior if required.