Add `riscv32-none-elf-gcc` as a possible RISC-V compiler
Nix uses `riscv32-none-elf` as the configuration name for the embedded
RISC-V 32-bit toolchain; check for this name in addition to
`riscv32-unknown-elf` and `riscv32-unknown-elf`.

It could be possible to add `riscv64-none-elf` too, but that would
require working multilib support in the toolchain, which is not the case
for the Nix-provided toolchain at the moment.