Add initial support for ADMH AD64
This PCB was found in a “Bongo Cat” keyboard from Taobao.  The stock
firmware identifies itself as “ADMH AD64”; the PCB has “Lm x AD design”
written on the top side, and no text on the bottom side.

This is a temporary firmware version with several known problems:

  - The USB IDs match the original firmware, but they are not even
    remotely unique (in particular, the vendor ID is 0xFEED).

  - The firmware uses RGB Lighting, even though the PCB actually has
    per-key RGB LEDs and therefore should use RGB Matrix.

The `default` and `via` keymaps were copied verbatim from the
`melgeek/mj64` board.  The value of `RGBLIGHT_LIMIT_VAL` was copied from
`kprepublic/bm60rgb` (also a 60% board with per-key addressable LEDs).