[keymap] xd87:sigprof-test: Add dynamic macro support
The same keys as on ID80 are used:

  - Fn+T — play macro 1;
  - Fn+G — play macro 2;
  - Fn+LCtrl+T — start recording macro 1;
  - Fn+LCtrl+G — start recording macro 2;
  - Fn+LCtrl+Esc — stop recording.

One difference from ID80 is that XD87 also has a right Ctrl key (on ID80
the rightmost modifier key is used as Fn instead of Ctrl); however,
supporting the usage of the right Ctrl key for dynamic macro control is
problematic, because the current method (`LM(_FN_CTL, MOD_LCTL)`) does
not support right side modifiers.  For now this problem is left
unresolved (macro recording is rare enough that a slight inconvenience
of not being able to use the right Ctrl key could be ignored).