[keymap] xd87:sigprof-test: Adjust backlight PWM frequency
Setting the backlight PWM frequency to 2000 Hz results in annoying
flicker, apparently because of interference with 1000 Hz USB interrupts.
Finding a good value seems to be hard, but using 0x1FFF for the PWM
resolution (which results in ≈1953 Hz PWM frequency) seems to work well
enough (some flicker can still be noticed at low brightness, but that's
probably unavoidable).