pintester: Add support for SWD pins on F103 bluepill
STM32F103 requires special manipulations with `AFIO->MAPR` to switch SWD
pins to the GPIO mode.  However, doing that unconditionally is not
desired, because those chips may need SWD for reflashing the bootloader,
and having SWD disabled while the firmware is running may make that
process more difficult.  Because of those concerns the SWD pins are
ignored by default, and the `AFIO->MAPR` change to turn SWD off is
performed only if at least one of those SWD pins is not ignored.