pintester: Add startup delay
If some keys appear as always pressed for any reason, the pintester code
might end up not reporting their state to the host, because it sends its
reports only when the matrix state changes, and if that happens before
the host completes USB enumeration and HID keyboard driver init, those
reports will be lost.  To avoid this information loss, add a delay
between the keyboard startup and the actual start of matrix scanning, so
that any matrix scan results would hopefully appear when the host is
ready to accept the generated key events.

The delay is 2000 ms by default, and can be changed by defining
`PINTESTER_STARTUP_DELAY` to the desired value in ms.