pintester: Add support for explicitly setting the input and output pins
By default all GPIO pins are tested as both inputs and outputs, so that
the matrix with any diode direction can be scanned.  However, in some
cases the input and output pins may be known beforehand, and specifying
the pin lists explicitly can result in more useful behavior when the
matrix does not contain diodes (instead of reporting two keypresses in
both directions at once, only one keypress will be reported).

The lists of input and output pins can be passed by defining the
`PINTESTER_INPUT_PINS` and `PINTESTER_OUTPUT_PINS` macros.  If one of
these macros is defined, the other one must be defined too, and
`PINTESTER_IGNORE_PINS` must not be defined (the default list of ignored
pins won't be used too).  An easy way to define those macros without
creating a `config.h` file is to pass variables with the same name to
`make` (or use the `-e` option of `qmk compile` or `qmk flash`).