Fix analog.c build when a converter is enabled
If a converter is enabled, the usual QMK pin names like `A0` are
reassigned accordingly to the pin names of the source MCU, and therefore
they cannot be used in the `pinToMux()` code.  Use ChibiOS names like
`PAL_LINE(GPIOA, 0)` in that code.

The conversion was done by a simple text replacement:

    '<,'>s/case \([A-Z]\)\([0-9]\+\):/case PAL_LINE(GPIO\1, \2):/