onekey: Fix console output on AVR
Plain `printf()` does not work for console output on AVR (the `xprintf`
code does not replace `printf()`, and QMK does not set up `stdout` for
the avr-libc `printf` implementation).  Use `uprintf()` instead, which
should work for all supported platforms.

Note that `uprintf()` requires the format string to be a constant (on
AVR it uses `PSTR()` internally to put the format string into PROGMEM),
therefore `uprintf(buffer)` cannot be used (and it would be wrong anyway
if the buffer has some chance to contain any format characters).