joystick: Add the JOYSTICK_SHARED_INPUT_PINS option
The analog joystick code performed some pin manipulations that might be
needed if the analog input pins are shared with something else, but
should not be needed if dedicated analog input pins are used.  Put that
code inside `#if defined(JOYSTICK_SHARED_INPUT_PINS)` blocks, so that
the normal case with non-shared pins is not slowed down by those pin
manipulations.

This change also makes it possible to connect capacitors to joystick
pins, so that they would work as hardware low pass filters and enable
using lower ADC sampling times.

The `setPinInput()` part is additionally enabled on AVR, because the
`analogReadPin()` implementation for AVR does not do that itself.