Remove forced release of keys on the one shot layer
The implementation of one shot layers forced the key from the one shot
layer to be released immediately after handling the press event (except
for some special cases like modifier keys).  That code was added in the
very first version of OSL implementation for QMK, when it was actually
needed, because in that version of QMK the `STRICT_LAYER_RELEASE` option
was still the default (or, more precisely, the `PREVENT_STUCK_MODIFIERS`
option, which existed at that time, was not the default).  However, in
more recent QMK version the source layers cache is enabled by default,
therefore the forced release of the key on the one shot layer is not
required (the normal release will be handled properly, because the layer
that was active at the press time is remembered).  And that forced
release actually breaks some use cases, because holding a key on some
one shot layer is not possible.

Disable the forced release of keys on the one shot layer unless either
the `STRICT_LAYER_RELEASE` or the `LEGACY_OSL_RELEASE` macros are
defined (in the `STRICT_LAYER_RELEASE` case the forced release is
actually needed to handle the release properly; the `LEGACY_OSL_RELEASE`
option is added just in case someone really wants to use the old broken
behavior even without `STRICT_LAYER_RELEASE`).

After this change it is possible to hold a key on some one shot layer
without having that event revert to a tap (the one shot layer still gets
turned off immediately after pressing that key, but the source layers
cache ensures that the key release would be handled using the same
mapping as was used for the press).