[keymap] id80:sigprof-test: Map Caps Lock tap to Esc
After getting a 65% keyboard I found that the best place for the Esc key
seems to be the Caps Lock location (this mapping avoids messing with the
grave/tilde key, which is mapped to a letter in the Russian keyboard
layout).  Reconfigure the ID80 keymap to map Caps Lock tap to Esc too.
The original Caps Lock key is available as Fn+Tab.