[keymap] id80:sigprof-test: Add BS_TOGG to the Adjust layer
My 65% board has a HHKB-style 1.5u Backspace at the backslash key
location; try to use the same location for Backspace on ID80 too (even
though using a 2u key for backslash won't be optimal).  This mapping is
not configured as default yet, and needs to be enabled by the magic key
(but the remapping persists in the EEPROM).