id80:sigprof-test: Actually swap Backspace with backslash in the keymap
Just using `BS_TOGG` to swap Backspace with backslash does not fully
produce the desired effect, because the backslash key was also
overloaded with a layer tap; although the tap function of that key still
changes as expected, the hold function still remains and does not move
to the new backslash location; this also interferes with autorepeat for
backspace, which is sometimes desired.  Apparently the only way to
achieve the desired behavior is to actually swap the keycodes in the
keymap (`BS_TOGG` can still be used to move them to the “standard”
position, although with the same problem for that backspace location).