[QP] Extract common code for surface backed drivers
Some drivers may need to keep a backing surface in RAM (e.g., because
the device does not provide a way to perform partial picture updates
with 1px granularity, and performing a read-modify-write operation is
also impossible or really slow).  Extract the code to deal with the
backing surface from the SH1106 OLED driver, so that it is not coupled
to the particular data formats and commands used by those OLEDs and can
be reused by other drivers.